refactor: entity 정리
