+++ domain/src/main/kotlin/org/monsing/course/CourseTicket.kt
@@ -5,7 +5,7 @@ import jakarta.persistence.Entity
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import org.monsing.BaseEntity
-import org.monsing.member.Teacher
+import org.monsing.member.teacher.Teacher
 
 @Entity
 class CourseTicket(
