--- domain/src/main/kotlin/org/monsing/member/CareerRepository.kt
@@ -1,7 +1,0 @@
-package org.monsing.member
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface CareerRepository : JpaRepository<Career, Long>
