+++ domain/src/main/kotlin/org/monsing/member/Review.kt
@@ -4,6 +4,7 @@ import jakarta.persistence.Entity
 import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import org.monsing.BaseEntity
+import org.monsing.member.teacher.Teacher
 
 @Entity
 class Review(
