--- domain/src/main/kotlin/org/monsing/member/TeacherRepository.kt
@@ -1,7 +1,0 @@
-package org.monsing.member
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface TeacherRepository : JpaRepository<Teacher, Long>
