+++ domain/src/main/kotlin/org/monsing/member/teacher/Career.kt
@@ -1,0 +1,12 @@
+package org.monsing.member.teacher
+
+import jakarta.persistence.Entity
+import org.monsing.BaseEntity
+
+@Entity
+class Career(
+
+    val period: String,
+
+    val detail: String
+) : BaseEntity()
