+++ domain/src/main/kotlin/org/monsing/member/teacher/Portfolio.kt
@@ -1,0 +1,11 @@
+package org.monsing.member.teacher
+
+import jakarta.persistence.Entity
+import org.monsing.BaseEntity
+import java.net.URL
+
+@Entity
+class Portfolio(
+
+    val url: URL
+) : BaseEntity()
