+++ domain/src/main/kotlin/org/monsing/member/teacher/Teacher.kt
@@ -1,11 +1,14 @@
-package org.monsing.member
+package org.monsing.member.teacher
 
 import jakarta.persistence.Entity
 import jakarta.persistence.EnumType
 import jakarta.persistence.Enumerated
 import jakarta.persistence.JoinColumn
+import jakarta.persistence.OneToMany
 import jakarta.persistence.OneToOne
 import org.monsing.BaseEntity
+import org.monsing.member.Member
+import org.monsing.member.StrongSideType
 
 @Entity
 class Teacher(
@@ -20,5 +23,14 @@ class Teacher(
 
     val description: String,
 
-    val forStudent: String
+    val forStudent: String,
+
+    val verified: Boolean,
+
+    @OneToMany
+    val portfolios: List<Portfolio> = mutableListOf(),
+
+    @OneToMany
+    val careers: List<Career> = mutableListOf()
+
 ) : BaseEntity()
